requirements.txt
absl-py==0.13.0
addict==2.4.0
albumentations==1.0.0
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
cycler==0.10.0
Cython==0.29.24
filelock==3.0.12
google-auth==1.35.0
google-auth-oauthlib==0.4.5
grpcio==1.39.0
huggingface-hub==0.0.15
idna==3.2
imageio==2.9.0
imgaug
importlib-metadata==4.6.4
joblib==1.0.1
kiwisolver==1.3.1
loguru==0.5.3
Markdown==3.3.4
matplotlib==3.4.3
networkx==2.6.2
numpy==1.21.2
oauthlib==3.1.1
opencv-python==4.2.0.32
opencv-python-headless==4.4.0.46
packaging==21.0
pandas==1.2.4
Pillow==8.3.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
regex==2021.8.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.45
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.1
Shapely==1.7.1
six==1.16.0
tabulate==0.8.9
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
thop==0.0.31.post2005241907
threadpoolctl==2.2.0
tifffile==2021.8.8
timm==0.6.7
tokenizers==0.10.3
tqdm==4.62.1
transformers @ git+https://github.com/huggingface/transformers@6f14eab50b702953d0e602339e16c7adee316004
typing-extensions==3.10.0.0
urllib3==1.26.6
Werkzeug==2.0.1
zipp==3.5.0
torchlibrosa==0.0.9
xgboost==1.6.2
